+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.7.1 - 2019-04-11
+### Fixed
+- If clicking box fails, give it one more try before moving on (#38)[https://github.com/jpchip/giveaway-grabber/issues/38]
+- Catch navigation errors (#48)[https://github.com/jpchip/giveaway-grabber/issues/48]
+- Check if on switch account page in more places (#49)[https://github.com/jpchip/giveaway-grabber/issues/49]
+
+
 ## 2.7.0 - 2019-03-31
 ### Fixed
 - Blacklist matches whole words (#43)[https://github.com/jpchip/giveaway-grabber/issues/43]